Add "setup_fiscal_calendar" function.
This makes it easier to change the global Calendar Settings without having
to jump through hoops.

It also makes it possible to simplify the fiscal_calendar implementation.

Note 1: The advanced docs will need to be updated.
Note 2: In order to be able to reliable test "validate_fiscal_calendar_params()"
        I had to add several tests classes. I added one class per commit
        to make inspection easier.

Fixes #3